+++ packages/generator-adp/src/app/types.ts
@@ -129,3 +129,19 @@ export type AttributePromptOptions = Partial<{
     [attributePromptNames.ui5ValidationCli]: CliValidationPromptOptions;
     [attributePromptNames.enableTypeScript]: EnableTypeScriptPromptOptions;
 }>;
+
+/**
+ * An interface representing the json input used to store the complete adaptation project
+ * generator configurations. The json is passed as an CLI argument.
+ */
+export interface JsonInput {
+    system: string;
+    client?: string;
+    username?: string;
+    password?: string;
+    application: string;
+    applicationTitle?: string;
+    targetFolder?: string;
+    projectName?: string;
+    namespace?: string;
+}